+++ src/backend/compare/report.ts
@@ -6,7 +6,7 @@ import type { DatabaseConfig } from '../db/types.js';
 import { Database } from '../db/db.js';
 import { assert, log } from '../logging.js';
 import { completeRequest, startRequest } from '../../perf-tracker.js';
-import { robustPath, siteConfig } from '../../util.js';
+import { robustPath, siteConfig } from '../util.js';
 import type { CompareView } from '../../views/view-types.js';
 import { prepareCompareView } from './prep-data.js';
