+++ src/backend/db/db.ts
@@ -19,7 +19,7 @@ import {
   FullPlotData,
   TimelineSuite
 } from '../../shared/api.js';
-import { robustPath, TotalCriterion } from '../../util.js';
+import { robustPath, TotalCriterion } from '../util.js';
 import { assert } from '../logging.js';
 import { simplifyCmdline } from '../../views/util.js';
 import type { SummaryStatistics } from '../../stats.js';
