+++ src/backend/dev-server/server.ts
@@ -2,7 +2,7 @@ import { ParameterizedContext } from 'koa';
 import { readFileSync } from 'node:fs';
 
 import { log } from '../logging.js';
-import { robustPath, robustSrcPath } from '../../util.js';
+import { robustPath, robustSrcPath } from '../util.js';
 
 export async function serveStaticResource(
   ctx: ParameterizedContext
