+++ src/backend/logging.ts
@@ -1,5 +1,5 @@
 import { Logger } from 'tslog';
-import { DEV, isRunningTests } from '../util.js';
+import { DEV, isRunningTests } from './util.js';
 
 // 0: silly, 1: trace, 2: debug, 3: info, 4: warn, 5: error, 6: fatal
 const trace = 2;
