+++ src/backend/project/data-export.ts
@@ -1,6 +1,6 @@
 import { existsSync } from 'node:fs';
 import { completeRequest, startRequest } from '../../perf-tracker.js';
-import { siteConfig, storeJsonGzip } from '../../util.js';
+import { siteConfig, storeJsonGzip } from '../util.js';
 import { log } from '../logging.js';
 import { Database } from '../db/db.js';
 import { ParameterizedContext } from 'koa';
