+++ src/timeline-calc.ts
@@ -1,6 +1,6 @@
 import { Database } from './backend/db/db.js';
 import { SummaryStatistics } from './stats.js';
-import { robustSrcPath } from './util.js';
+import { robustSrcPath } from './backend/util.js';
 import { Worker } from 'node:worker_threads';
 import { completeRequest, startRequest } from './perf-tracker.js';
 import { log } from './backend/logging.js';
